Restructure serialization APIs
