+++ package.json
@@ -23,7 +23,7 @@
     "prepare": "npm run build",
     "build": "tsc --build",
     "build:watch": "tsc --build --watch",
-    "build:docs": "typedoc src/index.doc.ts --plugin typedoc-github-theme",
+    "build:docs": "typedoc src --plugin typedoc-github-theme",
     "test": "node --test",
     "lint": "eslint src --ext .ts",
     "lint:fix": "eslint src --ext .ts --fix",
