--- src/index.doc.ts
@@ -1,1 +1,0 @@
-export * from "./index";
