+++ src/index.ts
@@ -1,5 +1,4 @@
 export * from "./common";
 export * from "./service";
 export * from "./handler";
-
-export { LazyValue, type Content, type Serializer } from "./serializer";
+export * from "./serialization";
